Use the first dimension when creating hypertable
`hypertable.dimensions` returns an instance of `ActiveRecord::Associations::CollectionProxy`, but it looks like we were expecting it to be just a single `Timescale::Dimension` instance. This causes the `timescale_hypertable` method to fail.

Using `hypertable.dimensions.first` allows `timescale_hypertable` to function correctly; it's possible that we'd want to process the additional dimensions and output `add_dimension` commands to the `schema.rb` file too, but I haven't tackled that here.