Add support for optional hypertable arguments
The current hypertable options only allow for a small subset of the
arguments supported by the plugin. It's relatively straightforward to
add generic support for _any_ hypertable options to be passed through.

As part of this effort, I've improved the SchemaDumper to handle the
majority of persistent options that impact the final database schema.